Implements ScaleCircles class to render a visual aid for map scale
Understanding the map scale with zoomable map is sometimes difficult.
Scale circles can help the user to estimate the distances.

This commit has a working example. The performance is sufficient on most
devices, but the oldest mobile devices may suffer from a slow rendering
due to DOM redraw. That's to be optimised in the future.